build: add find_package support, improve add_subdirectory behavior
