+++ include/SGP4/DecayedException.h
