+++ include/SGP4/Globals.h
