+++ include/SGP4/Observer.h
