+++ include/SGP4/SatelliteException.h
