+++ include/SGP4/SolarPosition.h
