+++ include/SGP4/TimeSpan.h
