+++ include/SGP4/TleException.h
