+++ include/SGP4/Vector.h
