--- libsgp4/DecayedException.cc
@@ -1,1 +1,0 @@
-#include "DecayedException.h"
