--- libsgp4/SatelliteException.cc
@@ -1,1 +1,0 @@
-#include "SatelliteException.h"
