--- libsgp4/TleException.cc
@@ -1,1 +1,0 @@
-#include "TleException.h"
