+++ src/DateTime.cpp
@@ -15,7 +15,7 @@
  */
 
 
-#include "DateTime.h"
+#include <SGP4/DateTime.h>
 
 #if 0
