+++ src/SatelliteException.cpp
@@ -1,0 +1,1 @@
+#include <SGP4/SatelliteException.h>
