+++ src/Util.cpp
@@ -15,4 +15,29 @@
  */
 
 
-#include "Vector.h"
+#include <SGP4/Util.h>
+
+#include <algorithm>
+#include <locale>
+#include <functional>
+
+namespace libsgp4::Util
+{
+    void TrimLeft(std::string& s)
+    {
+        s.erase(s.begin(),
+                std::find_if(s.begin(), s.end(), [](unsigned char c){ return std::isgraph(c) != 0; }));
+    }
+
+    void TrimRight(std::string& s)
+    {
+        s.erase(std::find_if(s.rbegin(), s.rend(), [](unsigned char c){ return std::isgraph(c) != 0; }).base(),
+                s.end());
+    }
+
+    void Trim(std::string& s)
+    {
+        TrimLeft(s);
+        TrimRight(s);
+    }
+} // namespace libsgp4::Util
